requirements.txt
# Core dependencies
opencv-python==4.8.1.*
numpy==1.24.*
Pillow==10.0.*

# Screenshot capture
pyautogui==0.9.54
selenium==4.15.*
webdriver-manager==4.0.*

# Image processing and OCR
pytesseract==0.3.10
scikit-image==0.22.*

# GUI (optional, for overlay)
PyQt5==5.15.*

# Utilities
python-dotenv==1.0.*
colorlog==6.7.*

# Development dependencies
pytest==7.4.*
black==23.11.*
flake8==6.1.*
